refactor: rename TimetableTheme's file to theme
