+++ lib/src/components/month_widget.dart
@@ -2,7 +2,7 @@ import 'package:black_hole_flutter/black_hole_flutter.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_layout_grid/flutter_layout_grid.dart';
 
-import '../styling.dart';
+import '../theme.dart';
 import '../utils.dart';
 import 'date_indicator.dart';
 import 'week_indicator.dart';
