+++ lib/src/components/now_indicator.dart
@@ -4,7 +4,7 @@ import 'package:black_hole_flutter/black_hole_flutter.dart';
 import 'package:flutter/material.dart';
 
 import '../date/controller.dart';
-import '../styling.dart';
+import '../theme.dart';
 import '../utils.dart';
 
 class NowIndicator extends StatefulWidget {
