+++ lib/src/components/time_indicators.dart
@@ -4,7 +4,7 @@ import 'package:black_hole_flutter/black_hole_flutter.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 
-import '../styling.dart';
+import '../theme.dart';
 import '../utils.dart';
 import 'time_indicator.dart';
